discord_dashboard/requirements.txt
Django==5.0.1
discord.py[voice]==2.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Pillow==10.0.1
asyncio-mqtt==0.13.0
openai>=1.54.0
PyNaCl>=1.5.0
pydub>=0.25.1
ffmpeg-python>=0.2.0


